requirement.txt
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
boto3==1.26.122
botocore==1.29.122
brotlipy==0.7.0
decorator==5.1.1
einops==0.6.1
encodec==0.1.1
executing==1.2.0
fastapi==0.95.1
filelock==3.12.0
fsspec==2023.4.0
funcy==2.0
h11==0.14.0
huggingface-hub==0.14.1
ipython==8.12.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.1
lit==16.0.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
netaddr==0.8.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.3.23
s3transfer==0.6.0
scipy==1.10.1
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
suno-bark @ git+https://github.com/suno-ai/bark.git@2c12023eb22868a633b76357b69d657b374736d9
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
uvicorn==0.22.0
wcwidth==0.2.6
WebOb==1.8.7
